Add config to refocus on vim after forward search
This is a hack suggested in https://ejmastnak.com/tutorials/vim-latex/pdf-reader/ to make the focus of the x server return to vim after a forward search in compiled pdf file. This configuration saved the active id of the vim window, and then use autocommand with vimtex event to call the external utility xdotool whenever a forward search is done.