Set path to Python3 executable for vim
Previously, vim and nvim inferred the executable for Python3 (needed in
various plugins) by relying on environment variables. This is however
unreliable in some cases, including using nvim as scroll back reader in
kitty. Therefore, this commit states explicitly what Python3 executable
to use and remove the reliance on environment variables for this
specific scenario.